refactor: add module architecture
