+++ src/constants.js
@@ -1,3 +1,3 @@
-export const ENDPOINT_DISCOVER = `${process.env.REACT_APP_API_URL}`;
-export const ENDPOINT_SEARCH = `${process.env.REACT_APP_API_URL}`;
+export const ENDPOINT_POPULAR_MOVIES = `${process.env.REACT_APP_API_URL}`;
+export const ENDPOINT_SEARCH_MOVIES = `${process.env.REACT_APP_API_URL}`;
 export const ENDPOINT_MOVIE = `${process.env.REACT_APP_API_URL}`;
