+++ src/modules/header/header.scss
@@ -25,11 +25,11 @@ header {
     max-width: 400px;
     padding: 0.5rem 1rem;
     border-radius: 1rem;
-    outline: 2px solid #ccc;
+    outline: none;
 
 
     &:focus {
-      outline: 2px solid black;
+      outline: 2px solid #ccc;
     }
   }
 }
