+++ src/modules/starred-movies/starred.scss
@@ -1,0 +1,1 @@
+// render star in all items in this view (all movies)
