+++ src/modules/watch-later/watchLater.test.js
@@ -1,7 +1,7 @@
 import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "./utils";
-import App from "../App";
+import { renderWithProviders } from "../../shared/utils/renderWithProviders";
+import App from "../../App";
 
 it("Watch Later movies page", async () => {
   renderWithProviders(<App />);
