+++ src/shared/components/movieCard/MovieCard.jsx
