+++ src/shared/components/movieGrid/movie-grid.scss
