+++ src/shared/utils/renderWithProviders.js
@@ -4,9 +4,9 @@ import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/dist/query";
-import moviesSlice from "../data/moviesSlice";
+import moviesSlice from "../../data/moviesSlice";
 import starredSlice from "../data/starredSlice";
-import watchLaterSlice from "../data/watchLaterSlice";
+import watchLaterSlice from "../../modules/watch-later/watchLaterSlice";
 
 export function renderWithProviders(
   ui,
